New database method to add skipped tracks (#15)
Polling service now adding skipped tracks to database
Added/updated unit tests
Updated datatbase row to bigint
Updated database model to long

+++ api/SpotifyPlaylistJanitorAPI/DataAccess/Entities/SkippedTrack.cs
@@ -6,7 +6,7 @@ public partial class SkippedTrack
 {
     public int Id { get; set; }
 
-    public int SkippedDate { get; set; }
+    public long SkippedDate { get; set; }
 
     public string SpotifyPlaylistId { get; set; } = null!;
